refactor: extract agent creator config and unify field add
